requirements.txt
mrcrowbar==0.9.0
pyserial==3.5
